requirements.txt
python-telegram-bot==20.0
onnxruntime
Pillow
numpy
python-dotenv